requirements.txt
flask==3.0.3
pillow==10.3.0
reportlab==4.2.0
sendgrid==6.11.0
requests==2.32.3
gunicorn==22.0.0
